Add centralized validation for firewall and daemon configurations
- Introduced `validate` package for centralized configuration validation.
- Added validation for paths, filenames, and specific extensions (e.g., `.toml`, `.pid`, `.log`).
- Enhanced `firewall` settings with validation logic (e.g., `table_name`, `chain_input_name`, `saves_rules_path`, `icmp_in_rate`, `input_drop`).
- Updated `daemon` PID file handling for additional checks on file type and content.
- Improved error handling and validation coverage for settings initialization.

+++ internal/cmd/daemon/start.go
@@ -54,7 +54,7 @@ func runDaemon(ctx context.Context, _ *cli.Command) error {
 		return err
 	}
 
-	err = d.Run(ctx, setting.Config.Testing, setting.Config.TestingInterval)
+	err = d.Run(ctx, setting.Config.Testing, uint16(setting.Config.TestingInterval))
 	if err != nil {
 		logger.Fatal(err.Error())
